Move lifting to original module
